--- a/lib/active_record/connection_adapters/oracle_enhanced/structure_dump.rb
+++ b/lib/active_record/connection_adapters/oracle_enhanced/structure_dump.rb
@@ -9,7 +9,8 @@ module ActiveRecord #:nodoc:
 
         def structure_dump #:nodoc:
           sequences = select(<<~SQL.squish, "SCHEMA")
-            SELECT sequence_name, min_value, max_value, increment_by, order_flag, cycle_flag
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */
+            sequence_name, min_value, max_value, increment_by, order_flag, cycle_flag
             FROM all_sequences
             where sequence_owner = SYS_CONTEXT('userenv', 'current_schema') ORDER BY 1
           SQL
@@ -18,7 +19,7 @@ module ActiveRecord #:nodoc:
             "CREATE SEQUENCE #{quote_table_name(result["sequence_name"])} MINVALUE #{result["min_value"]} MAXVALUE #{result["max_value"]} INCREMENT BY #{result["increment_by"]} #{result["order_flag"] == 'Y' ? "ORDER" : "NOORDER"} #{result["cycle_flag"] == 'Y' ? "CYCLE" : "NOCYCLE"}"
           end
           tables = select_values(<<~SQL.squish, "SCHEMA")
-            SELECT table_name FROM all_tables t
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ table_name FROM all_tables t
             WHERE owner = SYS_CONTEXT('userenv', 'current_schema') AND secondary = 'N'
             AND NOT EXISTS (SELECT mv.mview_name FROM all_mviews mv
                             WHERE mv.owner = t.owner AND mv.mview_name = t.table_name)
@@ -30,7 +31,7 @@ module ActiveRecord #:nodoc:
             virtual_columns = virtual_columns_for(table_name) if supports_virtual_columns?
             ddl = +"CREATE#{ ' GLOBAL TEMPORARY' if temporary_table?(table_name)} TABLE \"#{table_name}\" (\n"
             columns = select_all(<<~SQL.squish, "SCHEMA")
-              SELECT column_name, data_type, data_length, char_used, char_length,
+              SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ column_name, data_type, data_length, char_used, char_length,
               data_precision, data_scale, data_default, nullable
               FROM all_tab_columns
               WHERE table_name = '#{table_name}'
@@ -90,7 +91,7 @@ module ActiveRecord #:nodoc:
         def structure_dump_primary_key(table) #:nodoc:
           opts = { name: "", cols: [] }
           pks = select_all(<<~SQL.squish, "SCHEMA")
-            SELECT a.constraint_name, a.column_name, a.position
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ a.constraint_name, a.column_name, a.position
               FROM all_cons_columns a
               JOIN all_constraints c
                 ON a.constraint_name = c.constraint_name
@@ -109,7 +110,7 @@ module ActiveRecord #:nodoc:
         def structure_dump_unique_keys(table) #:nodoc:
           keys = {}
           uks = select_all(<<~SQL.squish, "SCHEMA")
-            SELECT a.constraint_name, a.column_name, a.position
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ a.constraint_name, a.column_name, a.position
               FROM all_cons_columns a
               JOIN all_constraints c
                 ON a.constraint_name = c.constraint_name
@@ -145,7 +146,7 @@ module ActiveRecord #:nodoc:
 
         def structure_dump_fk_constraints #:nodoc:
           foreign_keys = select_all(<<~SQL.squish, "SCHEMA")
-            SELECT table_name FROM all_tables
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ table_name FROM all_tables
             WHERE owner = SYS_CONTEXT('userenv', 'current_schema') ORDER BY 1
           SQL
           fks = foreign_keys.map do |table|
@@ -173,7 +174,7 @@ module ActiveRecord #:nodoc:
         def structure_dump_column_comments(table_name)
           comments = []
           columns = select_values(<<~SQL.squish, "SCHEMA")
-            SELECT column_name FROM user_tab_columns
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ column_name FROM user_tab_columns
             WHERE table_name = '#{table_name}' ORDER BY column_id
           SQL
 
@@ -207,7 +208,7 @@ module ActiveRecord #:nodoc:
         def structure_dump_db_stored_code #:nodoc:
           structure = []
           all_source = select_all(<<~SQL.squish, "SCHEMA")
-            SELECT DISTINCT name, type
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ DISTINCT name, type
             FROM all_source
             WHERE type IN ('PROCEDURE', 'PACKAGE', 'PACKAGE BODY', 'FUNCTION', 'TRIGGER', 'TYPE')
             AND name NOT LIKE 'BIN$%'
@@ -216,7 +217,7 @@ module ActiveRecord #:nodoc:
           all_source.each do |source|
             ddl = +"CREATE OR REPLACE   \n"
             texts = select_all(<<~SQL.squish, "all source at structure dump")
-              SELECT text
+              SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ text
               FROM all_source
               WHERE name = '#{source['name']}'
               AND type = '#{source['type']}'
@@ -239,7 +240,7 @@ module ActiveRecord #:nodoc:
         def structure_dump_views #:nodoc:
           structure = []
           views = select_all(<<~SQL.squish, "SCHEMA")
-            SELECT view_name, text FROM all_views
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ view_name, text FROM all_views
             WHERE owner = SYS_CONTEXT('userenv', 'current_schema') ORDER BY view_name ASC
           SQL
           views.each do |view|
@@ -251,7 +252,7 @@ module ActiveRecord #:nodoc:
         def structure_dump_synonyms #:nodoc:
           structure = []
           synonyms = select_all(<<~SQL.squish, "SCHEMA")
-            SELECT owner, synonym_name, table_name, table_owner
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ owner, synonym_name, table_name, table_owner
             FROM all_synonyms
             WHERE owner = SYS_CONTEXT('userenv', 'current_schema')
           SQL
@@ -264,13 +265,14 @@ module ActiveRecord #:nodoc:
 
         def structure_drop #:nodoc:
           sequences = select_values(<<~SQL.squish, "SCHEMA")
-            SELECT sequence_name FROM all_sequences where sequence_owner = SYS_CONTEXT('userenv', 'current_schema') ORDER BY 1
+            SELECT/*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */
+            sequence_name FROM all_sequences where sequence_owner = SYS_CONTEXT('userenv', 'current_schema') ORDER BY 1
           SQL
           statements = sequences.map do |seq|
             "DROP SEQUENCE \"#{seq}\""
           end
           tables = select_values(<<~SQL.squish, "SCHEMA")
-            SELECT table_name from all_tables t
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ table_name from all_tables t
             WHERE owner = SYS_CONTEXT('userenv', 'current_schema') AND secondary = 'N'
             AND NOT EXISTS (SELECT mv.mview_name FROM all_mviews mv
                             WHERE mv.owner = t.owner AND mv.mview_name = t.table_name)
@@ -286,7 +288,7 @@ module ActiveRecord #:nodoc:
 
         def temp_table_drop #:nodoc:
           temporary_tables = select_values(<<~SQL.squish, "SCHEMA")
-            SELECT table_name FROM all_tables
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ table_name FROM all_tables
             WHERE owner = SYS_CONTEXT('userenv', 'current_schema')
             AND secondary = 'N' AND temporary = 'Y' ORDER BY 1
           SQL
@@ -320,7 +322,7 @@ module ActiveRecord #:nodoc:
         # return [{'column_name' => 'FOOS', 'data_default' => '...'}, ...]
         def virtual_columns_for(table)
           select_all(<<~SQL.squish, "SCHEMA")
-            SELECT column_name, data_default
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ column_name, data_default
             FROM all_tab_cols
             WHERE virtual_column = 'YES'
             AND owner = SYS_CONTEXT('userenv', 'current_schema')
@@ -331,7 +333,7 @@ module ActiveRecord #:nodoc:
         def drop_sql_for_feature(type)
           short_type = type == "materialized view" ? "mview" : type
           features = select_values(<<~SQL.squish, "SCHEMA")
-            SELECT #{short_type}_name FROM all_#{short_type.tableize}
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ #{short_type}_name FROM all_#{short_type.tableize}
             where owner = SYS_CONTEXT('userenv', 'current_schema')
           SQL
           statements = features.map do |name|
@@ -342,7 +344,7 @@ module ActiveRecord #:nodoc:
 
         def drop_sql_for_object(type)
           objects = select_values(<<~SQL.squish, "SCHEMA")
-            SELECT object_name FROM all_objects
+            SELECT /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ object_name FROM all_objects
             WHERE object_type = '#{type.upcase}' and owner = SYS_CONTEXT('userenv', 'current_schema')
           SQL
           statements = objects.map do |name|