Add /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.2') */ hint to address slow SCHEMA queries

Migrating Oracle Database version to higher like 12c, I have experienced
some slowness mainly due to SCHEMA queries. This commit should address
this slowness.

I have tested multiple values of optimizer_features_enable from 19.1.0 to 11.2.0.2,
then there is a significant elapsed time of `bundle exec rspec` between
11.2.0.3 to 11.2.0.2.

- optimizer_features_enable = 11.2.0.3
Finished in 848 minutes 28 seconds (files took 0.61873 seconds to load)

- optimizer_features_enable = 11.2.0.2
Finished in 2 minutes 30 seconds (files took 0.61398 seconds to load)

### Environment
Oracle Database 19c (19.3) on Docker
Intel NUC Kit NUC6i5SYH with 4 CPU and 32 GB memory

This initialization parameter value can be changed to instance level
by running `alter system set optimizer_features_enable = '11.2.0.2'`,
which could affect other queries' execution plans.

Refer "1.232 OPTIMIZER_FEATURES_ENABLE"
https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/OPTIMIZER_FEATURES_ENABLE.html#GUID-E193EC9E-B642-4C01-99EC-24E04AEA1A2C