@koic 2020-10-14
Thank you for investigation and opening the PR. In summary, this patch improves performance to a proprietary Rails application.

This is the result of `time bin/rails db:migrate` (or `time bin/rails db:schema:dump`). Oracle version is `Oracle Database 19c Enterprise Edition Release 19.0.0.0.0`.

Below is the execution time without this patch.

```console
bin/rails db:migrate  4.87s user 2.07s system 0% cpu 18:14.28 total
```

Below is the execution time with this patch. So, this patch improves about 4~5x.

```console
bin/rails db:migrate  5.00s user 2.12s system 2% cpu 4:39.88 total
```

Below is the result of using a monkey patch by `OPTIMIZER_FEATURES_ENABLE` in a proprietary Rails application.

```diff
             result = select_all(<<~SQL.squish, "indexes", [bind_string("table_name", table_name)])
               SELECT
+                /*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.1') */
                 LOWER(i.table_name) AS table_name, LOWER(i.index_name) AS index_name, i.uniqueness,
                 i.index_type, i.ityp_owner, i.ityp_name, i.parameters,
                 LOWER(i.tablespace_name) AS tablespace_name,
```

It's almost no different from this patch by `OPT_PARAM`.

```console
bin/rails db:migrate  4.78s user 1.80s system 2% cpu 3:54.13 total
```

It was more faster in the days of Oracle 11.2.0-1.0 XE. So, performance is down due to the upgrade to Oracle 12c.

Please let me know if you have any necessary information. Thank you very much for your taking time.