@koic 2020-10-16
I previously told you the wrong information in an OSS online meetup.
The following is the result of gathering statistics in Oracle 19c.

```console
> sqlplus xxx/xxx@//x.x.x.x:xxxx/PDB1

SQL*Plus: Release 19.0.0.0.0 - Production on Fri Oct 16 10:05:00 2020
Version 19.3.0.0.0

Copyright (c) 1982, 2019, Oracle.  All rights reserved.

Last Successful login time: Fri Jun 05 2020 15:08:50 +09:00

Connected to:
Oracle Database 19c Enterprise Edition Release 19.0.0.0.0 - Production
Version 19.3.0.0.0

SQL> BEGIN
  2  DBMS_STATS.GATHER_DICTIONARY_STATS();
  3  END;
  4  /

PL/SQL procedure successfully completed.

SQL> BEGIN
  2  DBMS_STATS.GATHER_FIXED_OBJECTS_STATS;
  3  END;
  4  /

PL/SQL procedure successfully completed.

(snip)

% time bin/rails db:schema:dump
bin/rails db:schema:dump  4.70s user 1.92s system 2% cpu 4:11.56 total
```

So, it seems to get the same result as when using the hint phrase (`/*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.1') */`).

I hope it will be a clue to the performance regression for Oracle 12c...