@yahonda 2020-10-16
Thanks for the update. Let me summarize the elapsed time for each change.

* With `/*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.1') */` hint / Without `GATHER_DICTIONARY_STATS/GATHER_FIXED_OBJECTS_STATS` result is `bin/rails db:migrate  4.78s user 1.80s system 2% cpu 3:54.13 total`

* Without `/*+ OPTIMIZER_FEATURES_ENABLE('11.2.0.1') */` hint / With `GATHER_DICTIONARY_STATS/GATHER_FIXED_OBJECTS_STATS` result is `bin/rails db:schema:dump  4.70s user 1.92s system 2% cpu 4:11.56 total`