@yahonda 2020-10-20
@koic Would you test your `bin/rails db:schema:dump` against this pull request including https://github.com/rsim/oracle-enhanced/pull/2055/commits/4442257063419753c5accdb7ee644b342e7fb3b7 ?