@koic 2020-10-21
>  your `bin/rails db:schema:dump` against this pull request including 4442257 ?

I ran `bin/rails db:schema:dump` twice including the patch. It was still about 4-5 minutes.

```console
% time bin/rails db:schema:dump
bin/rails db:schema:dump  4.41s user 1.48s system 3% cpu 2:53.41 total

% time bin/rails db:schema:dump
bin/rails db:schema:dump  5.01s user 2.30s system 2% cpu 5:54.40 total
```