@yahonda 2020-10-27
Hi, Would you provide this information?

- Expected elapsed time like executed with Oracle database 11g.
- Provide Rails log file during `bin/rails db:schema:dump` which includes SCHEMA log entries.
- (If possible) Would you provide sql statement execution plans for "slow" sql statements?

If we need to talk offline, let's discuss it on Nov 5th.