@koic 2020-10-28
Yeah, let me have the discussion on Nov 5th. Thank you.